[fix]: Solve module importation error
